backend/app/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.11
h11==0.14.0
idna==3.10
iso8601==2.1.0
pillow==11.2.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pypika-tortoise==0.5.0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.4.0
pytz==2025.1
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
tortoise-orm==0.24.2
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.0
